+++ app/controllers/edge/internal/edge_hosts_controller.rb
@@ -43,8 +43,9 @@ class EdgeHostsController < RestController
         logger.debug(LogMessages::Endpoints::EndpointFinishedSuccessfully.new(
           "all_hosts:count replication for edge '#{Edge.get_name_by_hostname(current_user.role_id)}'"))
       else
+        sync_time = Time.now
         results = replicate_hosts(scope)
-        render(json: { "hosts": results })
+        render(json: { "hosts": results, "timestamp": sync_time})
         logger.debug(LogMessages::Endpoints::EndpointFinishedSuccessfullyWithLimitAndOffset.new(
           "all_hosts replication for edge '#{Edge.get_name_by_hostname(current_user.role_id)}'",
           limit,